requirements.txt
DateTime==4.8
decorator==5.1.1
deta==1.1.0
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
html5lib==1.1
keras==2.11.0
matplotlib==3.6.2
numpy==1.24.0
mplfinance==0.12.9b7
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.5.2
plotly==5.11.0
streamlit_option_menu==0.3.2
streamlit_lottie==0.0.3
streamlit-authenticator==0.2.1
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
streamlit==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
yahoo-fin==0.8.9.1
yfinance==0.2.3
python-telegram-bot==13.7